+++ source/DSCResources/MSFT_xRDSessionCollectionConfiguration/MSFT_xRDSessionCollectionConfiguration.psm1
@@ -398,4 +398,3 @@ function Test-TargetResource
 }
 
 Export-ModuleMember -Function *-TargetResource
-