+++ source/Examples/Resources/xRDSessionDeployment/1-CreateSessionDeployment.ps1
@@ -1,0 +1,18 @@
+<#
+    .DESCRIPTION
+        This example shows how to ensure a session deployment is created.
+#>
+
+Configuration Example
+{
+    Import-DscResource -ModuleName 'xRemoteDesktopSessionHost'
+
+    Node localhost {
+
+        xRDSessionDeployment RDSDeployment {
+            SessionHost = 'rdsessionhost.server.fqdn'
+            ConnectionBroker = 'connectionbroker.server.fqdn'
+            WebAccessServer =  'webaccess.server.fqdn'
+        }
+    }
+}